add release-monitoring.org fallback lookup method
The release-monitoring.org API provides version information for open
source packages. While this information can be out of date or
incorrect, it appears reliable enough to use as a
method-of-last-resort.

This change adds a release-monitoring.org fallback if no other version
lookup methods work. The lookup is done via a static list of project
IDs that I've assembled.

Signed-off-by: Steven Danna <steve@chef.io>