Fix gtk2 parse error
On download.gnome.org, the 2.x versions of gtk are in the gtk+
folder. Trying to look for the version information in the non-existent
gtk2 folder resulted in a parse error. Since the 2.x and 3.x versions
are in the same folder, I've constrained this to look at versions that
start with `2.[2-8]`. There is a 2.9x version, but from what I can see
other distributions are not shipping this and from some brief searches
it appears that it might have been the beta versions of gtk 3.

Fixes #4

Signed-off-by: Steven Danna <steve@chef.io>